[PostGIS] Cleaned warnings in md5.cpp.
git-svn-id: http://svn.osgeo.org/fdo/trunk@3903 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/PostGIS/Src/Provider/md5.cpp
@@ -137,7 +137,8 @@ void md5::init()
 
 void md5::update(const char* a_str)
 {
-    update(a_str, strlen(a_str));  // Optimization possible but not worth it.
+    // Optimization possible but not worth it.
+    update(a_str, static_cast<uint32_t>(strlen(a_str)));
 }
 
 void md5::update(const void* a_data, uint32_t a_data_size)
